+++ inc/common/Stream/ArrayStream.h
@@ -1,5 +1,5 @@
 /*
- * Stream implementation of a ring buffer.  The size is determined at instntiation,
+ * Stream implementation of a ring buffer.  The size is determined at instantiation,
  * and you can read / write to it just like any other stream.
  */
 
@@ -12,20 +12,20 @@ namespace digitalcave {
 	class ArrayStream : public Stream {
 
 		private:
-			uint8_t capacity;
+			uint16_t capacity;
 			volatile uint8_t* data;
-			volatile uint8_t head;
-			volatile uint8_t tail;
+			volatile uint16_t head;
+			volatile uint16_t tail;
 		public:
-			ArrayStream(uint8_t capacity);
+			ArrayStream(uint16_t capacity);
 			~ArrayStream();
 
 			uint8_t read(uint8_t *b);
 			uint8_t write(uint8_t b);
 
 			void clear();
-			uint8_t remaining();
-			uint8_t size();
+			uint16_t remaining();
+			uint16_t size();
 
 			uint8_t isEmpty();
 			uint8_t isFull();
